사용자 모드, 난이도 제외

+++ app/src/main/java/com/example/enejwl/dlthgud/Item.java
@@ -4,12 +4,12 @@ public class Item { // 아이템의 정보를 저장하는 클래스
     String name; // 아이템 이름
     int touch; // 아이템 클릭을 위한 터치 수
     int score; // 아이템 클릭 시 점수
-    int upMax; // 아이템이 올라와 있는 최대 시간
-    int upMin; // 아이템이 올라와 있는 최소 시간
+    double upMax; // 아이템이 올라와 있는 최대 시간
+    double upMin; // 아이템이 올라와 있는 최소 시간
     int image; // 화면에 나타날 이미지
     int period; //아이템 주기
 
-    public Item(String name, int touch, int score, int upMax, int upMin, int image, int period) {
+    public Item(String name, int touch, int score, double upMax, double upMin, int image, int period) {
         this.name = name;
         this.touch = touch;
         this.score = score;
@@ -43,19 +43,19 @@ public class Item { // 아이템의 정보를 저장하는 클래스
         return score;
     }
 
-    public void setUpMax(int upMax) {
+    public void setUpMax(double upMax) {
         this.upMax = upMax;
     }
 
-    public int getUpMax() {
+    public double getUpMax() {
         return upMax;
     }
 
-    public void setUpMin(int upMin) {
+    public void setUpMin(double upMin) {
         this.upMin = upMin;
     }
 
-    public int getUpMin() {
+    public double getUpMin() {
         return upMin;
     }
